imbd score

+++ etl/ddl_scripts_postgres/raw_data_vault/tables/rdv_film_sat.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS rdv.film_sat (
     , Film_Tijdsduur_min INTEGER NOT NULL
     , Genres_weging REAL NOT NULL
     , Film_tags VARCHAR NULL
-    , Film_IMDB_Score REAL NOT NULL
+    , Film_IMDB_Score NUMERIC(3,1) NOT NULL
     , Laaddatum TIMESTAMP NOT NULL
     , Is_Current INTEGER NOT NULL
     , PRIMARY KEY(Film_Hub_Key, Laaddatum)
